// Frontend/src/app/api/models/evaluation-response-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface EvaluationResponseDto {
  comment?: string;
  competenceScore?: number;
  createdAt?: string;
  diplomaScore?: number;
  evaluatorId?: number;
  evaluatorName?: string;
  evaluatorRole?: string;
  experienceScore?: number;
  id?: number;
  motivationScore?: number;
  softSkillsScore?: number;
  totalScore?: number;
  updatedAt?: string;
}