// Frontend/src/app/api/models/candidate-application-detail-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CandidateApplicationDetailDto {
  address?: string;
  cvUrl?: string;
  email?: string;
  firstName?: string;
  highestDegree?: string;
  id?: number;
  jobDescription?: string;
  jobOfferId?: number;
  jobTitle?: string;
  lastName?: string;
  lastUpdateDate?: string;
  specialization?: string;
  status?: string;
  submitDate?: string;
}